don't try to convert None data

+++ converter.py
@@ -107,9 +107,9 @@ for file_type, pattern, loader, processor, new_ext in (
         ))
 
         dump_name = get_dump_name(fname, new_ext)
-        jsonizer.dump(data, get_dump_name(fname, new_ext))
-        if dump_name != fname:
-            os.remove(fname)
-
-        n += 1
+        if data is not None:
+            jsonizer.dump(data, get_dump_name(fname, new_ext))
+            if dump_name != fname:
+                os.remove(fname)
+            n += 1
     logger.info('Converted {} {} files'.format(n, file_type))
